Drishti-Agent-main/backend/requirements.txt
# Project Drishti - Backend Dependencies
# Core FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Google Cloud & AI Services
google-cloud-aiplatform==1.43.0
google-cloud-vision==3.4.5
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
vertexai==1.43.0
firebase-admin==6.2.0

# Maps and Location Services
googlemaps==4.10.0
geopy==2.4.0

# Data Processing & ML
pandas==2.1.3
numpy==1.24.4
scikit-learn==1.3.2
opencv-python-headless==4.8.1.78
joblib==1.3.2

# Data Validation & Models
pydantic==2.5.0
pydantic[email]==2.5.0

# Image Processing
Pillow==10.1.0

# AI/ML
google-generativeai==0.3.2

# Async & HTTP
httpx==0.25.2
aiofiles==23.2.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0